Huge refactoring: Nls references are not nulls
To provide quick fixes for unresolved references UnresolvedReferenceQuickFixProvider can be used, but it works only for xml files.
Another variant is to bind IntentionAction to Annotator - for NlsFileImpl it works fine,
but since xml annotations done by IDEA default annotator this not good idea to reimplement this annotator.
The last variant(current) is to provide QuickFixes from Reference by implementing LocalQuickFixProvider.
But this works only when contributors contribute references always, even if it unresolvable - so contributors was fixed.
After this another problem was found for NlsNames in xml - completion provided from
DomConverter didn't used when reference is not null - in this case PsiReference.getVariants is used.